- Splitted out SDL2 surface output code builds now - Split SDL1 and SDL2 dependent output_surface.cpp parts into separate files to improve readability
